Remove OpCode.Field
no longer needed, we know fields in advance now via typeinfo
