+++ ulox/ulox.core/Package/Runtime/Enums/OpCode.cs
@@ -50,7 +50,6 @@
         GET_PROPERTY,
         SET_PROPERTY,
         METHOD,
-        FIELD,
         INVOKE,
 
         FREEZE,
